requirements.txt
astroid==2.8.0
attrs==19.3.0
autopep8==1.5.7
beautifulsoup4==4.11.1
bokeh==3.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docker==5.0.3
docutils==0.15.2
flake8==3.9.2
Flask==1.1.2
Flask-SocketIO==4.3.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
mccabe==0.6.1
mitmproxy==9.0.1
more-itertools==8.4.0
multipledispatch==0.6.0
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.9.0
platformdirs==2.3.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.7.0
pymongo==4.3.3
pyparsing==2.4.7
python-dotenv==1.0.0
python-dateutil==2.8.1
python-engineio==3.13.1
python-socketio==4.6.0
requests==2.24.0
six==1.15.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==1.2.3
typing-extensions==3.10.0.2
urllib3==1.25.9
wcwidth==0.2.5
websocket-client==1.2.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
